input axis int, must add input buffer

+++ Projects/Moving PlayerBox/Moving PlayerBox/GameData.h
@@ -5,6 +5,6 @@ namespace RB
 	class GameData
 	{
 	public:
-		float inputXAxis = 0.0f;
+		int inputXAxis = 0;
 	};
 }
